backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai
pydantic>=2.7,<3
httpx
pandas
numpy
scikit-learn
statsmodels
prophet
supabase>=2.4,<3
python-dotenv>=1.0,<2
